Rework logic, resolve some CodeClimate issues
